sample-django-app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
redis==5.0.1
django-prometheus==2.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9